feat: extract add resource into sub-command folder
