+++ packages/cli/src/commands/add/sub-commands/resource/create-resources.ts
